+++ packages/studio-base/src/panels/RawMessages/types.ts
@@ -1,0 +1,14 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+type AutoExpandMode = "auto" | "off" | "all";
+
+export type RawMessagesPanelConfig = {
+  autoExpandMode?: AutoExpandMode;
+  diffEnabled: boolean;
+  diffMethod: "custom" | "previous message";
+  diffTopicPath: string;
+  showFullMessageForDiff: boolean;
+  topicPath: string;
+};
